+++ NEWS.md
@@ -1,5 +1,7 @@
 # httr2 (development version)
 
+* New `check_resp_content_type()` to check response content types (#190).
+
 * Respect custom content type for `req_body_json()` (@mgirlich, #189).
 
 * `req_oauth_refresh()` now respects the `refresh_token` for caching (@mgirlich, #178).